requirements.txt
fastapi
uvicorn
streamlit
pymupdf
python-docx
groq
sqlmodel
python-dotenv
pandas